// TheSteelStoreCRM/src/app/core/models/steel-data.model.ts
export interface SteelDataModel {
  COILNUMBER: string;
  PRODUCT: string;
  LOCATION: string;
  QUALITY: string;
  THICKNESS_IN: string;
  THICKNESS_MM: string;
  WIDTH_IN: string;
  LENGTH_IN: string;
  diameter: string;
  WEIGHT_LB: string;
  GRADE: string;
  ROCKWELL: string;
  YIELD: string;
  ELONGATION: string;
  TENSILE: string;
  PASS_OIL: string;
  FINISH: string;
  TEMPER: string;
  COATING: string;
  PRODUCT_IMAGE: string;
  C: string;
  MN: string;
  P: string;
  S: string;
  SI: string;
  AL: string;
  CR: string;
  NB: string;
  TI: string;
  CA: string;
  N: string;
  NI: string;
  CU: string;
  V: string;
  B: string;
}